字典管理和字典项管理：增，删，改，查功能实现 。

+++ src/main/java/com/chiyun/outboundplatform/web/DictionaryListController.java
@@ -7,7 +7,11 @@ import com.chiyun.outboundplatform.service.IdictionaryListService;
 import com.chiyun.outboundplatform.service.IdictionaryService;
 import com.chiyun.outboundplatform.utils.StringUtil;
 import io.swagger.annotations.Api;
+import io.swagger.annotations.ApiImplicitParam;
+import io.swagger.annotations.ApiImplicitParams;
 import io.swagger.annotations.ApiOperation;
+
+import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RestController;
@@ -20,7 +24,7 @@ import java.util.List;
  */
 @Api(description = "字典项信息类型表")
 @RestController
-@RequestMapping(value = "/dict", method = {RequestMethod.GET, RequestMethod.POST})
+@RequestMapping(value = "/dict")
 public class DictionaryListController {
 
     @Resource
@@ -30,15 +34,20 @@ public class DictionaryListController {
     private IdictionaryListService idictionaryListService; //字典项
 
     @ApiOperation("查询所有字典项信息")
-    @RequestMapping("/findDictListAll")
+    @RequestMapping(value = "/findDictListAll",method = {RequestMethod.GET, RequestMethod.POST})
     public ApiResult<Object> findDictListAll(){
-        List<DictionaryListEntity> list= idictionaryListService.findAll();
-        return ApiResult.SUCCESS(list);
+
+            List<DictionaryListEntity> list = idictionaryListService.findAll();
+            return ApiResult.SUCCESS(list);
+
     }
 
-    @ApiOperation("添加一个字典项")
+
+
+    @ApiOperation("新增一个字典项")
     @RequestMapping(value ="/addDictList" ,method = RequestMethod.POST)
-    public ApiResult<Object> addDictList(DictionaryListEntity entity){
+    @ApiImplicitParam( name = "entity", value = "字典项对象", required = true, dataType = "DictionaryListEntity")
+    public ApiResult<Object> addDictList(@RequestBody DictionaryListEntity entity){
 
         String msg= pramesvalud(entity);
           if(msg.equals("ok")){
@@ -57,9 +66,15 @@ public class DictionaryListController {
 
     }
 
-    @ApiOperation("根据主键删除单个字典项")
+
+
+    @ApiOperation("根据【主键】删除单个字典项")
     @RequestMapping(value = "/deleteDictListById",method = RequestMethod.POST)
+    @ApiImplicitParam(paramType="query", name = "id", value = "主键", required = true, dataType = "Integer")
     public  ApiResult<Object>  deleteDictListById(Integer id){
+        if(id==0){
+            return ApiResult.FAILURE("删除失败,主键不能为空！！");
+        }
         int con=idictionaryListService.deleteById(id);
         if(con==1){
             return ApiResult.SUCCESS("删除成功");
@@ -69,10 +84,38 @@ public class DictionaryListController {
 
     }
 
-    @ApiOperation("根据字典英文名和字典项代号查询字典项的值")
-    @RequestMapping("/queryByZdywmAndKey")
-    public  ApiResult<Object> queryByZdywmAndKey(String zdywm, String key){
 
+    @ApiOperation("根据[字典名表主键]批量删除字典项")
+    @RequestMapping(value = "/deleteDictListByDid",method = RequestMethod.POST)
+    @ApiImplicitParam(paramType="query", name = "did", value = "外键", required = true, dataType = "Integer")
+      public  ApiResult<Object>  deleteDictListByDid(Integer did){
+
+        if(did==0){
+            return ApiResult.FAILURE("删除失败,主键不能为空！！");
+        }
+        int con=idictionaryListService.deleteByDid(did);
+        if(con>=0){
+            return ApiResult.SUCCESS("删除成功,共删除了:"+con+"个字典项");
+        }else{
+            return ApiResult.FAILURE("删除失败");
+        }
+      }
+
+
+
+    @ApiOperation("根据[字典英文名和字典项代号]查询字典项的值")
+    @RequestMapping(value="/queryDictListByZdywmAndKey",method = {RequestMethod.GET, RequestMethod.POST})
+    @ApiImplicitParams({
+            @ApiImplicitParam(paramType="query", name = "zdywm", value = "字典英文名", required = true, dataType = "String"),
+          @ApiImplicitParam(paramType ="query",name = "key",value = "字典项代码",required = true,dataType = "String")
+    })
+    public  ApiResult<Object> queryDictListByZdywmAndKey(String zdywm, String key){
+        if(StringUtil.isNull(zdywm)){
+            return  ApiResult.FAILURE("(zdywm):字典英文名不能为空!!!");
+        }
+        if(StringUtil.isNull(key)){
+            return  ApiResult.FAILURE("(key):字典项代码不能为空!!!");
+        }
         String  value= idictionaryListService.querDictListByZdywmAndKey(zdywm,key);
         if(StringUtil.isNull(value)){
             return  ApiResult.FAILURE("查询失败，未查询到对应的值");
@@ -80,9 +123,21 @@ public class DictionaryListController {
         return  ApiResult.SUCCESS(value);
     }
 
-    @ApiOperation("根据字典中文名和字典项代号查询字典项的值")
-    @RequestMapping("/queryByZdzwmAndKey")
-    public  ApiResult<Object> queryByZdzwmAndKey(String zdzwm, String key){
+
+
+    @ApiOperation("根据[字典中文名和字典项代号]查询字典项的值")
+    @RequestMapping(value = "/queryDictListByZdzwmAndKey",method = {RequestMethod.GET, RequestMethod.POST})
+    @ApiImplicitParams({
+            @ApiImplicitParam(paramType="query", name = "zdzwm", value = "字典中文名", required = true, dataType = "String"),
+            @ApiImplicitParam(paramType ="query",name = "key",value = "字典项代码",required = true,dataType = "String")
+    })
+    public  ApiResult<Object> queryDictListByZdzwmAndKey(String zdzwm, String key){
+        if(StringUtil.isNull(zdzwm)){
+            return  ApiResult.FAILURE("(zdywm):字典中文名不能为空!!!");
+        }
+        if(StringUtil.isNull(key)){
+            return  ApiResult.FAILURE("(key):字典项代码不能为空!!!");
+        }
         String  value= idictionaryListService.querDictListByZdzwmAndKey(zdzwm,key);
         if(StringUtil.isNull(value)){
             return  ApiResult.FAILURE("查询失败，未查询到对应的值");
@@ -90,7 +145,146 @@ public class DictionaryListController {
         return  ApiResult.SUCCESS(value);
     }
 
-     //不能通过其他地方获取到的不为空的字段校验
+
+
+    @ApiOperation("更新一个字典项")
+    @RequestMapping(value = "/updateDictList",method = RequestMethod.POST)
+    @ApiImplicitParam( name = "entity", value = "字典项对象", required = true, dataType = "DictionaryListEntity")
+    public ApiResult<Object> updateDictList (@RequestBody DictionaryListEntity entity){
+
+           if(entity.getId()==0){
+               return  ApiResult.FAILURE("主键不能为空");
+           }
+        entity= DicListEntuty(entity) ;  //信息补全
+        int  cn=idictionaryListService.updateOne(entity);
+        if(cn==1){
+            return ApiResult.SUCCESS("更新成功");
+        } else if (cn==0){
+            return ApiResult.FAILURE("更新成功,但0条记录被更新了");
+        }else if(cn>1){
+            return ApiResult.FAILURE("更新失败,但"+cn+"条记录被更新");
+        }
+        else{
+            return  ApiResult.FAILURE("更新失败");
+        }
+
+
+    }
+
+
+
+    @ApiOperation("根据[字典名表主键]注销多个关联的字典项")
+    @RequestMapping(value = "/zhuxiaoDicListByDid",method = RequestMethod.POST)
+    @ApiImplicitParam(paramType="query", name = "did", value = "外键", required = true, dataType = "Integer")
+    public  ApiResult<Object>  zhuxiaoDicListByDid(Integer did){
+             if(did==0){
+                 return   ApiResult.FAILURE("did 外键不能为空");
+             }
+            int cn= idictionaryListService.zhuXiaoByDid(did);
+             if (cn>0){
+                 return  ApiResult.SUCCESS("注销成功，共注销了："+cn+"个字典项");
+             }
+             else{
+                 return  ApiResult.FAILURE("注销失败！！");
+             }
+
+
+    }
+
+
+    @ApiOperation("根据[主键]注销单个字典项")
+    @RequestMapping(value = "/zhuxiaoDicListById",method = RequestMethod.POST)
+    @ApiImplicitParam(paramType="query", name = "id", value = "主键", required = true, dataType = "Integer")
+    public  ApiResult<Object>   zhuxiaoDicListById(Integer id){
+        if(id==0){
+            return   ApiResult.FAILURE("did 外键不能为空");
+        }
+        int cn= idictionaryListService.zhuXiaoOne(id);
+        if (cn==1){
+            return  ApiResult.SUCCESS("注销成功");
+        } else if(cn==0){
+            return  ApiResult.FAILURE("注销失败,零条记录被注销");
+        } else if(cn>1){
+            return ApiResult.FAILURE("注销成功,但"+cn+"条记录被注销了");
+        }
+        else{
+            return  ApiResult.FAILURE("注销失败！！");
+        }
+
+    }
+
+
+
+    @ApiOperation("根据[主键]查询单个字典项")
+    @RequestMapping(value = "/findDictListById",method = RequestMethod.POST)
+    @ApiImplicitParam(paramType="query", name = "id", value = "主键", required = true, dataType = "Integer")
+     public ApiResult<Object>  findDictListById (Integer id){
+        if(id==0){
+            return   ApiResult.FAILURE("did 外键不能为空");
+        }
+
+            DictionaryListEntity entity = idictionaryListService.findById(id);
+        if(entity==null){
+            return  ApiResult.FAILURE("未查询到该字典项！");
+        }
+            return  ApiResult.SUCCESS(entity);
+
+
+
+
+
+    }
+
+
+    @ApiOperation("根据[字典名表主键]查询关联的多个字典项")
+    @RequestMapping(value = "/findDictListByDid",method = RequestMethod.POST)
+    @ApiImplicitParam(paramType="query", name = "did", value = "外键", required = true, dataType = "Integer")
+    public  ApiResult<Object>  findDictListByDid(Integer did){
+        if(did==0){
+            return   ApiResult.FAILURE("did 外键不能为空");
+        }
+        List<DictionaryListEntity> entitys = idictionaryListService.findBydid(did);
+        if(entitys==null||entitys.size()==0){
+            return  ApiResult.FAILURE("未查询到该字典项！");
+        }
+        return  ApiResult.SUCCESS(entitys);
+
+    }
+
+
+    @ApiOperation("根据[字典中文名]查询关联的多个字典项")
+    @RequestMapping(value = "/findDictListByZdzwm",method = RequestMethod.POST)
+    @ApiImplicitParam(paramType="query", name = "zdzwm", value = "字典中文名", required = true, dataType = "String")
+     public  ApiResult<Object>  findDictListByZdzwm(String  zdzwm){
+        if(StringUtil.isNull(zdzwm)){
+            return  ApiResult.FAILURE("字典中文名不能为空！！");
+        }
+         List<DictionaryListEntity>  entitys=idictionaryListService.findByZdzwm(zdzwm);
+         if(entitys==null||entitys.size()==0){
+             return  ApiResult.FAILURE("未查询到该字典项！");
+         }
+         return  ApiResult.SUCCESS(entitys);
+     }
+
+
+
+    @ApiOperation("根据[字典英文名]查询关联的多个字典项")
+    @RequestMapping(value = "/findDictListByZdywm",method = RequestMethod.POST)
+    @ApiImplicitParam(paramType="query", name = "zdywm", value = "字典英文名", required = true, dataType = "String")
+    public  ApiResult<Object>  findDictListByZdywm(String  zdywm){
+        if(StringUtil.isNull(zdywm)){
+            return  ApiResult.FAILURE("字典中文名不能为空！！");
+        }
+        List<DictionaryListEntity>  entitys=idictionaryListService.findByZdywm(zdywm);
+        if(entitys==null||entitys.size()==0){
+            return  ApiResult.FAILURE("未查询到该字典项！");
+        }
+        return  ApiResult.SUCCESS(entitys);
+    }
+
+
+
+    //不能通过其他地方获取到的不为空的字段校验
     private  String   pramesvalud(DictionaryListEntity entity){
           String str=null;
           if( entity.getDid()==null||entity.getDid()==0){
@@ -106,7 +300,7 @@ public class DictionaryListController {
 
     }
 
-    // 避免某些属性为空，导致sql报错
+    // 避免某些属性为空，导致sql报错   新增的时候用到
     private DictionaryListEntity EntityisHebing(DictionaryListEntity fromEn){
 
         if(!StringUtil.isNull(fromEn.getZdzwm())&&!StringUtil.isNull(fromEn.getZdywm())){
@@ -120,8 +314,18 @@ public class DictionaryListController {
             }
             return  fromEn;
         }
-
-
     }
+        //更新的时候用到 单个
+        private    DictionaryListEntity DicListEntuty(DictionaryListEntity fromEn) {
+            DictionaryListEntity  queEn=idictionaryListService.findById(fromEn.getId());
+             if(StringUtil.isNull(fromEn.getZdzwm())){ fromEn.setZdzwm(queEn.getZdzwm()); }
+             if(StringUtil.isNull(fromEn.getZdywm())){fromEn.setZdywm(queEn.getZdywm());}
+             if(StringUtil.isNull(fromEn.getC_id())){fromEn.setC_id(queEn.getC_id());}
+             if(StringUtil.isNull(fromEn.getC_name())){fromEn.setC_name(queEn.getC_name());}
+             if(StringUtil.isNull(fromEn.getZxbz())){fromEn.setZxbz(queEn.getZxbz());}
+             if(fromEn.getDid()==0){fromEn.setDid(queEn.getDid());}
+            return  fromEn;
+        }
+
 
 }
